refactor: port 관련 코드 제거
